+++ CHANGELOG.md
@@ -22,9 +22,11 @@ Here is a template for new release sections
 ## [Unreleased]
 
 ### Added
--
+- Warning for missing parameter when parsing inputs from epa to mvs (#656)
+- New module `exceptions.py` in `multi_vector_simulator.utils` to gather custom MVS exceptions (#656)
 ### Changed
--
+- Function `utils.compare_input_parameters_with_reference` accepts parameters as dict for json comparison (#656)
+- Move A1 and C0 custom exceptions into `multi_vector_simulator.utils.exceptions.py` (#656)
 ### Removed
 -
 ### Fixed